US125094: Move generic _logTelemetryEvent into d2l-activity-editor-telemetry-mixin Have activity ActivityQuizEditorTelemtryMixin extend base activity telemetry mixin
