+++ components/d2l-activity-editor/mixins/d2l-activity-editor-telemetry-mixin.js
@@ -30,6 +30,22 @@ export const ActivityEditorTelemetryMixin = superclass => class extends supercla
 	_getSaveStartMarkName(type) {
 		return `d2l-activity-${type}-editor.page.save.start`;
 	}
+
+	async _logTelemetryEvent(href, action, type, telemetryId) {
+		if (!href || !action || !type || !telemetryId) return;
+
+		const eventBody = new Events.EventBody()
+			.setAction(action)
+			.setObject(href, type);
+		const event = new Events.TelemetryEvent()
+			.setType('TelemetryEvent')
+			.setDate(new Date())
+			.setSourceId(telemetryId)
+			.setBody(eventBody);
+		const client = await D2L.Telemetry.CreateClient();
+		client.logUserEvent(event);
+	}
+
 	async _logUserEvent(href, action, type, telemetryId, performanceMeasureName) {
 		if (!href || !action || !type || !telemetryId || !performanceMeasureName) return;
